Players and movement.
In this commit, I have added two players. I added a movement script allowing the players to accelerate and drift. They acceleration works by speeding up the player while the forward key is pressed up to a max and the drift works by slowly decreasing the speed and turning only spins the player not changing it's direction while not accelerating. Both players have both been added to a 'world' scene.