Added powerups
I have created another progress bar for each player that acts as the cooldown for the powerups. When this bar fulls the players can activate their powerups. So far I have added 3 powerups. 2x faster shooting, shot gun, and star shot. It is random what one of these is given to the player. This powerup is then activated for a set ammount of time/bullets.