Asteroids!!! - partly
Asteroids added. I need to get collition shapes for the 3 smaller asteroids and instance more asteroids into the world scene. Movement for ships has been reworked to allow bounce mechanics, 1/4 of health is removed when asteroid is hit. the asteroids go of a path from one side of the screen to another. ya :)