Added powerup indicators
I have added a small icon to show what powerup you could activate. This is done with a sprite sheet that swaps framed depending of what powerup the player has next. I have also changed the colours of the bullet particles making the base particle white and modulating it whether the bullet is blue or red.