Finished adding menu and fixed death bugs
I have finished adding functionallity to the menu. There is a quit button exits the game, a score display with a reset score button that sets these scores back to 0, 1 is added to the winning ships score when the other dies. A auto play option so the players wont have to go back to menue screen after every game(this is a switch). And a play button that starts the game. This screne has a title and a paralax star background. I have fixed powerups carring over after death.